Add postDraw and preUpdate signals.
The "postDraw" signal is dispatched after the current state has been drawn. The "preUpdate" signal is dispatched before Flixel updates the current state (and its children).

Those are the signals plugins should subscribe to in order to draw on the screen or update itself along with Flixel.

DebugPathDisplay and TimerManager have been refactored to rely on those signals.