Add show/hide/toggleVisibility and remove forceDebugger
The debugger initialization is now "lazy", which means Flixel will initiate a bare minimum debugger at first. If FlxG.debugger.show() is called, the debugger is properly initialized.

It's now possible to programmatically change the debugger visibility using show(), hide() and toggleVisibility().

Since the debugger is now lazyly initialized, I've removed the forceDebugger property in FlxGame. As discussed in #158, it's was causing confusion and problems. If FlxG.debug is false and a developer still wants to show the debugger, a call to FlxG.debugger.show(true) will force the initialization.