FlxSprite seems to be working.
This is not the best approach to achieve that, but I've tried to keep changes to a minimum. This commit also differs from my original idea of implementing a render that is completely decoupled from renderable entities (e.g. FlxSprite).

I've left some hardwired rendering code in FlxSprite (see calcFrame(), for instance), but they will be removed as I progress.