Cameras are working in GPU render (change breaks draw() signature).
I've tried my best to implement the GPU render without affecting the current drawing API, but that was not possible. The way Flixel draws entities is completely unfriendly for a GPU render using cameras. My best bet was to use render to texture to overcome the problems, but that didn't work either.

I was forced to change the draw() method to draw(c:Camera), which means the Flixel render will now invoke draw() for each entity, telling it to render to the specified camera.