Debug tiles in tilemaps with GPU render.
The code is extremely ugly and it needs some serious love. It's just a prof of concept for now, but it's not good. When showing debug tiles, the framerate dropped from 30 to 20 because the tilemap is being rendered twice.

I've left several TODOs and clues to refactor the code later.